Added cypress e2e testing framework, moved jest tests to component folders
