+++ .circleci/config.yml
@@ -43,13 +43,13 @@ jobs:
           key: v1-circleci-nextjs-project-{{ .Branch }}-{{ .Revision }}
       - run:
           name: Jest tests
-          command: cd build_path && npm run test:unit
+          command: cd build_path && npm run test:jest
       - store_artifacts:
           path: ~/project/build_path/src/tests/__image_snapshots__/__diff_output__
           destination: diff_output
       - run:
           name: Update image snapshots when tests fails
-          command: cd build_path && npm run test:unit -- -u
+          command: cd build_path && npm run test:jest:update
           when: on_fail
       - store_artifacts:
           path: ~/project/build_path/src/tests/__image_snapshots__
