+++ cypress/integration/home-page.spec.ts
@@ -1,0 +1,11 @@
+import {
+  smallBreakpoint,
+  mediumBreakpoint,
+  largeBreakpoint,
+} from "@/utils/constants";
+
+describe("The Home Page", () => {
+  it("successfully loads homepage", () => {
+    cy.visit("/");
+  });
+});
