+++ cypress/plugins/index.ts
@@ -1,0 +1,14 @@
+/// <reference types="cypress" />
+// const { initPlugin } = require("cypress-plugin-snapshots/plugin");
+
+// module.exports = (on, config) => {
+//   initPlugin(on, config);
+//   return config;
+// };
+
+// import { initPlugin } from "cypress-plugin-snapshots/plugin";
+
+// export default (on, config) => {
+//   initPlugin(on, config);
+//   return config;
+// };
