+++ jest.config.js
@@ -2,9 +2,14 @@ module.exports = {
   verbose: true,
   testEnvironment: "jsdom",
   setupFilesAfterEnv: ["<rootDir>/jest.setup.ts"],
-  testPathIgnorePatterns: ["<rootDir>/.next/", "<rootDir>/node_modules/"],
+  testPathIgnorePatterns: [
+    "<rootDir>/cypress/",
+    "<rootDir>/.next/",
+    "<rootDir>/node_modules/",
+  ],
   moduleNameMapper: {
     "\\.(scss|sass|css)$": "identity-obj-proxy",
     "@/components/(.*)": "<rootDir>/src/components/$1",
+    "@/utils/(.*)": "<rootDir>/src/utils/$1",
   },
 };
