+++ package.json
@@ -20,6 +20,7 @@
     "@types/react": "^17.0.8",
     "@typescript-eslint/eslint-plugin": "^4.28.1",
     "@typescript-eslint/parser": "^4.28.0",
+    "cypress": "^7.7.0",
     "eslint": "^7.27.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-jsx-a11y": "^6.4.1",
@@ -36,7 +37,11 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "test:unit": "jest",
+    "test:jest": "jest",
+    "test:jest:update": "jest -- -u",
+    "test:cypress:open": "cypress open",
+    "test:cypress:run": "cypress run",
+    "test:cypress:run:update": "cypress run --env updateSnapshots=true",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix"
   },
