+++ src/components/welcome/__image_snapshots__/welcome-test-tsx-should-match-image-snapshot-1-snap.png
