+++ src/utils/constants.ts
@@ -1,0 +1,3 @@
+export const smallBreakpoint = 640;
+export const mediumBreakpoint = 768;
+export const largeBreakpoint = 1024;
