+++ tsconfig.json
@@ -20,6 +20,7 @@
     "jsx": "preserve",
     "paths": {
       "@/components/*": ["components/*"],
+      "@/utils/*": ["utils/*"],
     }
   },
   "include": [
@@ -28,6 +29,8 @@
     "**/*.tsx"
   ],
   "exclude": [
-    "node_modules"
+    "node_modules",
+    "**/*.test.ts",
+    "**/*.test.tsx"
   ]
 }
